+++ src/Mount.tsx
@@ -1,5 +1,0 @@
-import { defineComponent } from 'vue-demi'
-
-export default defineComponent({
-	//
-})