+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change logs
 
+## 2023.02.15 v1.1.0
+
+1. Add support for customized tagName
+2. mount target allows passing in a string
+3. fix vNode.id name
+4. upgrade all package version
+
 ## 2023.02.06 v1.0.0
 
 1. create project